Begin migration to NSKeyedArchiver. Minor pokes.

+++ macintosh/sceneDocument/colorPicker.mm
@@ -154,7 +154,7 @@
 	// this doesn't work right now,
 	// formatter keeps formatting
 	// will look for a way to change color on the fly
-	int tag=[[aNotification object]tag];
+	NSInteger tag=[[aNotification object]tag];
 	switch (tag)
 	{
 		case cRedEdit:	//image with
